User Riv vim plugin from fork
fork is a PullRequest to Rykka/riv.vim that allows to configure
whether to use a `:fixdel` hack in Riv or not
(currenlty it is not used only on neovim).

Using `:fixdel` on all my setups (Linux + Gnome terminal + tmux)
breaks delete key when Riv is active.

Using fork until this PR109 is merged to riv.vim master.